added/restructured holiday stop flow steps
