+++ app/client/components/user.tsx
@@ -26,7 +26,10 @@ import {
   ConsentsBanner,
   SuppressConsentBanner
 } from "./consent/consentsBanner";
-import { HolidayStopFlow } from "./holiday/holidayStopFlow";
+import { HolidayConfirmed } from "./holiday/holidayConfirmed";
+import { HolidayDateChooser } from "./holiday/holidayDateChooser";
+import { HolidayReview } from "./holiday/holidayReview";
+import { HolidaysOverview } from "./holiday/holidaysOverview";
 import { EmailAndMarketing } from "./identity/EmailAndMarketing";
 import { Main } from "./main";
 import { MembershipFAQs } from "./membershipFAQs";
@@ -119,18 +122,30 @@ const User = () => (
       {Object.values(ProductTypes)
         .filter(shouldHaveHolidayStopsFlow)
         .map((productType: ProductType) => (
-          <HolidayStopFlow
+          <HolidaysOverview
             key={productType.urlPart}
             path={"/suspend/" + productType.urlPart}
             productType={productType}
-            currentStep={1}
+            currentStep={0}
           >
-            {/* <ConfirmDates
-              path="confirm"
+            <HolidayDateChooser
+              path="new"
               productType={productType}
-              currentStep={2}
-            /> */}
-          </HolidayStopFlow>
+              currentStep={1}
+            >
+              <HolidayReview
+                path="review"
+                productType={productType}
+                currentStep={2}
+              >
+                <HolidayConfirmed
+                  path="confirmed"
+                  productType={productType}
+                  currentStep={3}
+                />
+              </HolidayReview>
+            </HolidayDateChooser>
+          </HolidaysOverview>
         ))}
 
       <MembershipFAQs path="/help" />
